refactor: 리뷰 받은 사항 수정
- dto 패키지 수정
- supressWarning 관련 주석 추가
