+++ main/src/main/java/org/sopt/makers/crew/main/admin/v2/dto/PropertyResponse.java
@@ -2,7 +2,6 @@ package org.sopt.makers.crew.main.admin.v2.dto;
 
 import java.util.Map;
 
-import org.sopt.makers.crew.main.admin.v2.service.dto.PropertyDto;
 import org.sopt.makers.crew.main.entity.property.Property;
 
 public record PropertyResponse(String key, Map<String, Object> value) {
@@ -14,4 +13,5 @@ public record PropertyResponse(String key, Map<String, Object> value) {
 	public static PropertyResponse from(PropertyDto dto) {
 		return new PropertyResponse(dto.key(), dto.properties());
 	}
+
 }
