+++ main/src/main/java/org/sopt/makers/crew/main/admin/v2/service/AdminFacade.java
@@ -4,8 +4,7 @@ import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
 
-import org.sopt.makers.crew.main.admin.v2.service.dto.AdminPagePresenter;
-import org.sopt.makers.crew.main.admin.v2.service.dto.PropertyDto;
+import org.sopt.makers.crew.main.admin.v2.dto.AdminPagePresenter;
 import org.sopt.makers.crew.main.admin.v2.service.vo.PropertyVo;
 import org.sopt.makers.crew.main.entity.property.Property;
 import org.springframework.stereotype.Service;
@@ -31,10 +30,6 @@ public class AdminFacade {
 		Map<String, String> formattedJsonMap = jsonPrettierService.prettierJson(allProperties);
 
 		return AdminPagePresenter.create(allProperties, formattedJsonMap);
-	}
-
-	public PropertyDto findPropertyByKey(String key) {
-		return PropertyDto.from(adminService.findPropertyByKey(key));
 	}
 
 }
