+++ main/src/main/java/org/sopt/makers/crew/main/admin/v2/service/AdminService.java
@@ -3,7 +3,7 @@ package org.sopt.makers.crew.main.admin.v2.service;
 import java.util.List;
 import java.util.Map;
 
-import org.sopt.makers.crew.main.admin.v2.service.dto.HomeProperties;
+import org.sopt.makers.crew.main.admin.v2.dto.HomeProperties;
 import org.sopt.makers.crew.main.entity.property.Property;
 
 public interface AdminService {
